+++ src/features/oauth/index.ts
@@ -4,4 +4,4 @@ export * from './hooks/use-authorize-form';
 export * from './hooks/use-client';
 export * from './services/get-client';
 export * from './services/get-clients';
-export * from './services/use-consent';
+export * from './services/consent';
